+++ spec/spec_helper.rb
@@ -10,6 +10,9 @@ SimpleCov.start do
 end
 
 require "webmock/rspec"
+
+Dir[File.expand_path("support/**/*.rb", __dir__)].each { |f| require f }
+
 require "eussiror"
 
 RSpec.configure do |config|